requirements.txt
Flask==2.0.3
SQLAlchemy==2.0.21
psycopg2-binary
gunicorn
gspread
oauth2client
python-dotenv
Werkzeug<2.1.0



